+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 # Zappa Changelog
 
+## 0.43.2
+* Add index prefix to each subsequent schedule expression for the same event (#1051)
+* fix/Only update CORS on resources which have an OPTIONS method (#1036)
+* Support for binary support and cors simultaneously (#1011)
+* Set Flask script name based on domain which the request was made to (#1015)
+* Fix SNS Async (#1055)
+* GZip for slim handler
+* Force color option
+* Various dep bumps 
+
+## 0.43.1
+* Fixes #1001, don't override AWS env vars if k:v not set. Thanks Nik and Sean!
+
 ## 0.43.0
+* Checks for the key 'Environment' when fetching remote lambda env vars (#988)
+* except BotoCoreError before general exception in zappa.cli.update
+* make cookie hack case-insensitive
+* Fix #998 - Make environment variable keys strings instead of byte arrays in python 3.6
+* Add --disable_progress command line parameter
+* #946 - Allow setting cors to false.
+* #870 Lambda from outside
+* Implement context header mappings - Feature Request Issue #939
+* Separating out native AWS environment variables ##962
+* Rule name shortening
+* Splintering aws_environment_variables from environment_variables (to avoid overwriting AWS native env vars).
+
+## 0.42.2
 * Add exclude for __pycache__ contents (#943)
 * Fix #937 - Use get_data
 * Add support for configuring APIGW cache TTL and encryption #942